prisma/dataimport/requirements.txt
pandas=1.4.3
pyodbc=4.0.39
python-dotenv=1.0.0